+++ src/Resolver/Factory.php
@@ -11,6 +11,7 @@ use React\Dns\Protocol\Parser;
 use React\Dns\Protocol\BinaryDumper;
 use React\EventLoop\LoopInterface;
 use React\Dns\Query\RetryExecutor;
+use React\Dns\Query\TimeoutExecutor;
 
 class Factory
 {
@@ -36,7 +37,11 @@ class Factory
 
     protected function createExecutor(LoopInterface $loop)
     {
-        return new Executor($loop, new Parser(), new BinaryDumper());
+        return new TimeoutExecutor(
+            new Executor($loop, new Parser(), new BinaryDumper(), null),
+            5.0,
+            $loop
+        );
     }
 
     protected function createRetryExecutor(LoopInterface $loop)